# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_nabiilNajm26_go-bank_internal_domain.Account:
    properties:
      account_number:
        type: string
      account_type:
        $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AccountType'
      balance:
        type: number
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AccountStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.AccountStatus:
    enum:
    - active
    - inactive
    - frozen
    - closed
    type: string
    x-enum-varnames:
    - AccountStatusActive
    - AccountStatusInactive
    - AccountStatusFrozen
    - AccountStatusClosed
  github_com_nabiilNajm26_go-bank_internal_domain.AccountType:
    enum:
    - savings
    - checking
    - deposit
    type: string
    x-enum-varnames:
    - AccountTypeSavings
    - AccountTypeChecking
    - AccountTypeDeposit
  github_com_nabiilNajm26_go-bank_internal_domain.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.User'
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.CreateAccountRequest:
    properties:
      account_type:
        allOf:
        - $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AccountType'
        enum:
        - savings
        - checking
        - deposit
      currency:
        type: string
    required:
    - account_type
    - currency
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 255
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.UpdateAccountRequest:
    properties:
      account_type:
        allOf:
        - $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AccountType'
        enum:
        - savings
        - checking
        - deposit
      status:
        allOf:
        - $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AccountStatus'
        enum:
        - active
        - inactive
        - frozen
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 255
        minLength: 3
        type: string
      phone:
        type: string
    type: object
  github_com_nabiilNajm26_go-bank_internal_domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      phone:
        type: string
      profile_image_url:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Banking API with authentication, transfers, and caching
  title: GoBank API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new bank account for authenticated user
      parameters:
      - description: Account creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      description: Delete account if it has zero balance (safety measure)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update account name or other details for authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account details
      tags:
      - accounts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: User login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /users/profile:
    delete:
      description: Delete user account and all associated data (irreversible)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user profile
      tags:
      - users
    get:
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user profile information (name, email, phone)
      parameters:
      - description: User update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_nabiilNajm26_go-bank_internal_domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /users/profile/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload profile image to S3 with safety limits (5MB max, PNG/JPG
        only)
      parameters:
      - description: Profile image file (PNG, JPG, JPEG, max 5MB)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload profile image
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"